feat: variable utils module fix: regex at variable name validation
